ds1302, devsd: Split code for the DISCARD segment into separate files. Includes required platform changes for n8vem-mark4, p112.
