+++ Kernel/dev/devide_discard.c
@@ -18,9 +18,9 @@
 #include <blkdev.h>
 
 /****************************************************************************/
-/* Code below this point used only once, at startup, so we want it to live  */
+/* Code in this file is used only once, at startup, so we want it to live   */
 /* in the DISCARD segment. sdcc only allows us to specify one segment for   */
-/* each source file. This "solution" is a bit (well, very) hacky ...        */
+/* each source file.                                                        */
 /****************************************************************************/
 
 static void devide_delay(void)
