+++ Kernel/dev/devsd.c
@@ -10,6 +10,8 @@
 /* and http://elm-chan.org/docs/mmc/mmc_e.html                           */
 /*-----------------------------------------------------------------------*/
 
+#define _SD_PRIVATE
+
 #include <kernel.h>
 #include <kdata.h>
 #include <printf.h>
@@ -18,14 +20,7 @@
 #include <stdbool.h>
 #include <blkdev.h>
 
-/* internal functions */
-static void sd_init_drive(uint8_t drive);
-static int sd_spi_init(uint8_t drive);
-static void sd_spi_release(uint8_t drive);
-static int sd_send_command(uint8_t drive, unsigned char cmd, uint32_t arg);
-static uint8_t sd_spi_wait(uint8_t drive, bool want_ff);
-
-static uint8_t devsd_transfer_sector(void)
+uint8_t devsd_transfer_sector(void)
 {
     uint8_t attempt, drive;
     bool success;
@@ -66,13 +61,13 @@ static uint8_t devsd_transfer_sector(void)
     return 0;
 }
 
-static void sd_spi_release(uint8_t drive)
+void sd_spi_release(uint8_t drive)
 {
     sd_spi_raise_cs(drive);
     sd_spi_receive_byte(drive);
 }
 
-static uint8_t sd_spi_wait(uint8_t drive, bool want_ff)
+uint8_t sd_spi_wait(uint8_t drive, bool want_ff)
 {
     unsigned int timer;
     unsigned char b;
@@ -97,7 +92,7 @@ static uint8_t sd_spi_wait(uint8_t drive, bool want_ff)
     return b;
 }
 
-static int sd_send_command(uint8_t drive, unsigned char cmd, uint32_t arg)
+int sd_send_command(uint8_t drive, unsigned char cmd, uint32_t arg)
 {
     unsigned char n, res, *p;
 
@@ -144,118 +139,4 @@ static int sd_send_command(uint8_t drive, unsigned char cmd, uint32_t arg)
     }while ((res & 0x80) && --n);
 
     return res;         /* Return with the response value */
-}
-
-/****************************************************************************/
-/* Code below this point used only once, at startup, so we want it to live  */
-/* in the DISCARD segment. sdcc only allows us to specify one segment for   */
-/* each source file. This "solution" is a bit (well, very) hacky ...        */
-/****************************************************************************/
-
-DISCARDABLE
-
-void devsd_init(void)
-{
-    uint8_t d;
-
-    for(d=0; d<SD_DRIVE_COUNT; d++)
-        sd_init_drive(d);
-}
-
-static void sd_init_drive(uint8_t drive)
-{
-    blkdev_t *blk;
-    unsigned char csd[16], n;
-    uint8_t card_type;
-
-    kprintf("SD drive %d: ", drive);
-    card_type = sd_spi_init(drive);
-
-    if(!(card_type & (~CT_BLOCK))){
-        kprintf("no card found\n");
-        return;
-    }
-
-    blk = blkdev_alloc();
-    if(!blk)
-        return;
-
-    blk->transfer = devsd_transfer_sector;
-    blk->driver_data = (drive & DRIVE_NR_MASK) | card_type;
-    
-    /* read and compute card size */
-    if(sd_send_command(drive, CMD9, 0) == 0 && sd_spi_wait(drive, false) == 0xFE){
-        for(n=0; n<16; n++)
-            csd[n] = sd_spi_receive_byte(drive);
-        if ((csd[0] >> 6) == 1) {
-            /* SDC ver 2.00 */
-            blk->drive_lba_count = ((uint32_t)csd[9] 
-                                   + (uint32_t)((unsigned int)csd[8] << 8) + 1) << 10;
-        } else {
-            /* SDC ver 1.XX or MMC*/
-            n = (csd[5] & 15) + ((csd[10] & 128) >> 7) + ((csd[9] & 3) << 1) + 2;
-            blk->drive_lba_count = (csd[8] >> 6) + ((unsigned int)csd[7] << 2) 
-                                   + ((unsigned int)(csd[6] & 3) << 10) + 1;
-            blk->drive_lba_count <<= (n-9);
-        }
-    }
-    sd_spi_release(drive);
-
-    blkdev_scan(blk, 0);
-}
-
-static int sd_spi_init(uint8_t drive)
-{
-    unsigned char n, cmd, card_type, ocr[4];
-    timer_t timer;
-
-    sd_spi_raise_cs(drive);
-
-    sd_spi_clock(drive, false);
-    for (n = 20; n; n--)
-        sd_spi_receive_byte(drive); /* 160 dummy clocks */
-
-    card_type = CT_NONE;
-    /* Enter Idle state */
-    if (sd_send_command(drive, CMD0, 0) == 1) {
-        /* initialisation timeout 1 second */
-        timer = set_timer_sec(1);
-        if (sd_send_command(drive, CMD8, (uint32_t)0x1AA) == 1) {    /* SDHC */
-            /* Get trailing return value of R7 resp */
-            for (n = 0; n < 4; n++) ocr[n] = sd_spi_receive_byte(drive);
-            /* The card can work at vdd range of 2.7-3.6V */
-            if (ocr[2] == 0x01 && ocr[3] == 0xAA) {
-                /* Wait for leaving idle state (ACMD41 with HCS bit) */
-                while(!timer_expired(timer) && sd_send_command(drive, ACMD41, (uint32_t)1 << 30));
-                /* Check CCS bit in the OCR */
-                if (!timer_expired(timer) && sd_send_command(drive, CMD58, 0) == 0) {
-                    for (n = 0; n < 4; n++) ocr[n] = sd_spi_receive_byte(drive);
-                    card_type = (ocr[0] & 0x40) ? CT_SD2 | CT_BLOCK : CT_SD2;   /* SDv2 */
-                }
-            }
-        } else { /* SDSC or MMC */
-            if (sd_send_command(drive, ACMD41, 0) <= 1) {
-                /* SDv1 */
-                card_type = CT_SD1;
-                cmd = ACMD41;
-            } else {
-                /* MMCv3 */
-                card_type = CT_MMC;
-                cmd = CMD1;
-            }
-            /* Wait for leaving idle state */
-            while(!timer_expired(timer) && sd_send_command(drive, cmd, 0));
-            /* Set R/W block length to 512 */
-            if(timer_expired || sd_send_command(drive, CMD16, 512) != 0)
-                card_type = CT_NONE;
-        }
-    }
-    sd_spi_release(drive);
-
-    if (card_type) {
-        sd_spi_clock(drive, true); /* up to 20MHz should be OK */
-        return card_type;
-    }
-
-    return CT_NONE; /* failed */
 }
