+++ Kernel/dev/ds1302.h
@@ -8,6 +8,8 @@ uint8_t rtc_secs(void);
    http://datasheets.maximintegrated.com/en/ds/DS1302.pdf table 3 */
 void ds1302_read_clock(uint8_t *buffer, uint8_t length);
 
+uint8_t uint8_from_bcd(uint8_t value);
+
 /* platform code must provide these functions */
 void ds1302_set_pin_clk(bool state);
 void ds1302_set_pin_ce(bool state);
