+++ Kernel/platform-n8vem-mark4/fuzix.lnk
@@ -33,11 +33,14 @@ swap.rel
 devsys.rel
 platform-n8vem-mark4/devtty.rel
 platform-n8vem-mark4/devide.rel
+platform-n8vem-mark4/devide_discard.rel
 platform-n8vem-mark4/devsd.rel
+platform-n8vem-mark4/devsd_discard.rel
 platform-n8vem-mark4/devsdspi.rel
 platform-n8vem-mark4/mbr.rel
 platform-n8vem-mark4/blkdev.rel
 platform-n8vem-mark4/ds1302.rel
+platform-n8vem-mark4/ds1302_discard.rel
 platform-n8vem-mark4/ds1302-mark4.rel
 platform-n8vem-mark4/monitor.rel
 -e
