+++ Kernel/platform-p112/Makefile
@@ -1,12 +1,14 @@
 CSRCS += devices.c main.c devtty.c devfd.c
 DSRCS = ../dev/blkdev.c ../dev/devide.c ../dev/mbr.c ../dev/ds1302.c
+DDSRCS = ../dev/devide_discard.c ../dev/ds1302_discard.c
 ASRCS = crt0.s z180.s commonmem.s p112.s ds1302-p112.s devfd2.s monitor.s flopboot.s
 
 AOBJS = $(ASRCS:.s=.rel)
 COBJS = $(CSRCS:.c=.rel)
 DOBJS = $(patsubst ../dev/%.c,%.rel, $(DSRCS))
+DDOBJS = $(patsubst ../dev/%.c,%.rel, $(DDSRCS))
 
-OBJS  = $(AOBJS) $(COBJS) $(DOBJS)
+OBJS  = $(AOBJS) $(COBJS) $(DOBJS) $(DDOBJS)
 
 CROSS_CCOPTS += -I../dev/
 
@@ -22,6 +24,9 @@ $(COBJS): %.rel: %.c
 
 $(DOBJS): %.rel: ../dev/%.c
 	$(CROSS_CC) $(CROSS_CCOPTS) -c $<
+
+$(DDOBJS): %.rel: ../dev/%.c
+	$(CROSS_CC) $(CROSS_CCOPTS) $(CROSS_CC_SEGDISC) -c $<
 
 clean:
 	rm -f $(OBJS) $(JUNK)  core *~ fuzix.com fuzix-boot.fdd
