+++ Kernel/platform-p112/fuzix.lnk
@@ -33,9 +33,11 @@ swap.rel
 devsys.rel
 platform-p112/devtty.rel
 platform-p112/devide.rel
+platform-p112/devide_discard.rel
 platform-p112/blkdev.rel
 platform-p112/mbr.rel
 platform-p112/ds1302.rel
+platform-p112/ds1302_discard.rel
 platform-p112/ds1302-p112.rel
 platform-p112/monitor.rel
 platform-p112/devfd.rel
